requirements.txt
fastapi
uvicorn
alembic
sqlalchemy
psycopg2-binary
python-dotenv

